+++ src/index.js
@@ -22,7 +22,7 @@
  */
 
 /** Engine version string. */
-export const VERSION = '1.2.0';
+export const VERSION = '1.2.1';
 
 /* ========================================================================== *
  * MATH